requirements.txt
numpy
pandas
nltk
chromadb
sentence-transformers
tqdm
streamlit
google-generativeai